// src/types/models/MembershipVO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Gym Manager Swagger
 * Gym Manager REST API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface MembershipVO
 */
export interface MembershipVO {
    /**
     * 회원권 명
     * @type {string}
     * @memberof MembershipVO
     */
    membershipName?: string;
    /**
     * 회원권 기간
     * @type {string}
     * @memberof MembershipVO
     */
    membershipPeriod?: string;
    /**
     * 회원권 시퀀스
     * @type {number}
     * @memberof MembershipVO
     */
    membershipSeq?: number;
    /**
     * 등록 일자
     * @type {Date}
     * @memberof MembershipVO
     */
    regDate?: Date;
    /**
     * 등록자 ID
     * @type {string}
     * @memberof MembershipVO
     */
    regId?: string;
    /**
     * 사업장 시퀀스
     * @type {number}
     * @memberof MembershipVO
     */
    wpSeq?: number;
}

/**
 * Check if a given object implements the MembershipVO interface.
 */
export function instanceOfMembershipVO(value: object): value is MembershipVO {
    return true;
}

export function MembershipVOFromJSON(json: any): MembershipVO {
    return MembershipVOFromJSONTyped(json, false);
}

export function MembershipVOFromJSONTyped(json: any, ignoreDiscriminator: boolean): MembershipVO {
    if (json == null) {
        return json;
    }
    return {
        
        'membershipName': json['membershipName'] == null ? undefined : json['membershipName'],
        'membershipPeriod': json['membershipPeriod'] == null ? undefined : json['membershipPeriod'],
        'membershipSeq': json['membershipSeq'] == null ? undefined : json['membershipSeq'],
        'regDate': json['regDate'] == null ? undefined : (new Date(json['regDate'])),
        'regId': json['regId'] == null ? undefined : json['regId'],
        'wpSeq': json['wpSeq'] == null ? undefined : json['wpSeq'],
    };
}

export function MembershipVOToJSON(json: any): MembershipVO {
    return MembershipVOToJSONTyped(json, false);
}

export function MembershipVOToJSONTyped(value?: MembershipVO | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'membershipName': value['membershipName'],
        'membershipPeriod': value['membershipPeriod'],
        'membershipSeq': value['membershipSeq'],
        'regDate': value['regDate'] == null ? undefined : ((value['regDate']).toISOString()),
        'regId': value['regId'],
        'wpSeq': value['wpSeq'],
    };
}